Add day 7 part one solution
Ok I'll admit.. this solution only seems to work for the example
that was given in the README. I must have misunderstood some key
part of the instructions, but I just can't seem to pinpoint it.

Given that this is just for fun, I'm committing this as-is, with
the (noble) intention of coming back. I hope "future me" comes
back to this one day...