Fix erroneous logic with day7 solution
The previous solution was close, but not quite getting it. I thought
through a bit more, and decided there was likely something wrong
with the fundamental insertion of the nodes, and realized it would
likely be better if I stuck with something a little more.. primative.

As it turns out, the code is a lot more concise when using a map..
although I will full admit maps of maps are usually hateful, but
this wasn't so bad.