Add the day10 initial commit
This commit gives us a first crack at the solution.. but we still
have a probelm with making this work in all directions. Right now
we can grow our velocity in the positive X and positive Y axis, but
we cannot for negative numbers.. in order for this to work, I will
have to double back and fix it.